pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "concertvenues"
version = "0.1.0"
description = "Scrape upcoming concert events from multiple venues and generate a static website"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "jinja2>=3.1",
    "python-dateutil>=2.9",
    "playwright>=1.40",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "ruff>=0.4",
]

[project.scripts]
cv = "concertvenues.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["concertvenues*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
